refactor(server): extract LLM client → src/server/llm.js
Pure move, zero behavior change. Pulls the shared Anthropic client
singleton (`anthropic`, 20-minute timeout) and the `dateContext()`
prompt-date helper out of server.js into src/server/llm.js, then
re-imports both.

Why: continues the server.js dismemberment — isolating the LLM client
so the prompt/generation surface stops anchoring on the monolith.

What:
- New src/server/llm.js exports `anthropic` (new Anthropic with the
  20-min timeout) and `dateContext()` (verbatim).
- server.js imports `{ anthropic, dateContext }` from it; the
  module-level singleton + function bodies are replaced with breadcrumbs.
- KEPT the bare `import Anthropic from '@anthropic-ai/sdk'` — 4 handlers
  still spin up their own short-timeout `new Anthropic(...)` client
  locally (lines 6908, 7201, 17329, 17503). Those are untouched.

Test plan:
- `node --check server.js` → OK
- `npm run lint` (no-undef gate) → clean
- route-inventory guard → 213 routes PASS
- vitest → 41/41 pass

Rollback: revert this commit; the singleton + dateContext were a pure
copy, so reverting restores the inline definitions exactly.
